t/data/parsing/temma-node-1.dat

#data
<t:element/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma element>
|     <link>

#data
axs<t:element/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "axs"
|     <temma element>
|     <link>

#data
<t:element>a</t:elemenT><t:element/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma element>
|       "a"
|     <temma element>
|     <link>

#data
<t:element><p>foo</t:element><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma element>
|       <p>
|         "foo"
|     <link>

#data
<t:element><p>f</t:element><p>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma element>
|       <p>
|         "f"
|   <body>
|     <p>
|       "b"

#data
<t:attr/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <head>
|     <link>

#data
axs<t:attr/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "axs"
|     <temma attr>
|     <link>

#data
<t:attr>a</t:attr><t:attr/><link/>
#errors
1;19;m;unmatched end tag;;t:attr
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <body>
|     "a"
|     <temma attr>
|     <link>

#data
<t:attr><p>foo</t:attr><link/>
#errors
1;24;m;unmatched end tag;;t:attr
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <body>
|     <p>
|       "foo"
|       <link>

#data
<t:attr><p>f</t:attr><p>b
#errors
1;22;m;unmatched end tag;;t:attr
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <body>
|     <p>
|       "f"
|     <p>
|       "b"

#data
<meta><t:attr/><base/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <meta>
|     <temma attr>
|     <base>

#data
<meta><t:attr><link>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <meta>
|     <temma attr>
|     <link>

#data
<html><t:attr name='"manifest"' value='1243'/><head>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|     name=""manifest""
|     value="1243"
|   <head>

#data
<t:comment/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma comment>
|     <link>

#data
axs<t:comment/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "axs"
|     <temma comment>
|     <link>

#data
<t:comment>a</t:comment><t:comment/><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma comment>
|       "a"
|     <temma comment>
|     <link>

#data
<t:comment><p>foo</t:comment><link/>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma comment>
|       <p>
|         "foo"
|     <link>

#data
<t:comment><p>f</t:comment><p>b
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma comment>
|       <p>
|         "f"
|   <body>
|     <p>
|       "b"

#data
<svg><t:element><g/></t:element></svg>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     <svg svg>
|       <temma element>
|         <svg g>

#data
<t:value>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma value>

#data
a<t:value>bbb
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <temma value>
|     "bbb"

#data
a<t:value>b</t:value>bb
#errors
1;22;m;unmatched end tag;;t:value
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "a"
|     <temma value>
|     "bbb"

#data
<t:attr><t:value>bbb
#errors
#document
| <!DOCTYPE html>
| <html>
|   <temma attr>
|   <head>
|     <temma value>
|   <body>
|     "bbb"

#data
<t:value t:parse>hoge</t:value>xxy
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <temma value>
|       temma parse=""
|       "hoge"
|   <body>
|     "xxy"
